Add 64x64 Bits Polynomial Multiplication Using Linux x86 Assembly Code
Polynomial multiplication, also known as carriless multiplication, is
very important in many cryptography applications including AES-GCM.

This implements the 64x64 bits polynomial multiplication results in
128 bits result. This code is for Linux x86 (32 bits) build only.

This may not work if the CPU does not have support of the pclmulqdq
instruction. Check by cat /proc/cpuinfo and look for it in flags.