Add 64x64 Bits Polynomial Multiplication Using Aarch64 Code
Polynomial multiplication, also known as carriless multiplication, is
very important in many cryptography applications including AES-GCM.

This implements the 64x64 bits polynomial multiplication results in
128 bits result. On Armv7, the vmull.p8 instruction first proposed
by Conrado Gouvea. See references:
    https://conradoplg.cryptoland.net/files/2010/12/mocrysen13.pdf
    https://conradoplg.cryptoland.net/software/ecc-and-ae-for-arm-neon/

Fortunately starting with armv8, with the aarch64, there are two much
faster instructions to use, pmull and pmull1, which can do 64x64 to
128 bits polynomial multiplication in one instruction. I am using it.

To find out whether your linux arm platform supports the instruction,
just cat /proc/cpuinfo and look for flags or features and look for
the keyword pmull to see if it is listed or not. Any Aarch64 CPU
should have it.