Consolidate definition of u128 and speedup aes-gcm computation.
I consolidated the definition of struct u128 so it can be shared
by aes128. Also leverage polynomial production function to speed
up the aes-gcm computation. Please note, it is unfortunate that
in the official specification of AES-GCM, not only is the byte
order is big endian (first byte is the most significant one), in
contrary to little endian on most widely used x86 and ARM chips,
but also the most significant bit in a byte is the leading bit.
For example, if the first byte is 0x80, then bit 0 is 1 and there
are 7 0s following it. This causes confusion and frustration in
leveraging existing hardware instructions. See the discussion of
"Bit Reflection Peculiarity of GCM" in 2014 Intel White Paper:
  https://software.intel.com/sites/default/files/managed/72/cc/clmul-wp-rev-2.02-2014-04-20.pdf